cast enum to int to satisfy new gcc version
The latest gcc complains about this code with the following:
  180 |     if ((accessSetMode == doNotSet) || (accessSetMode == reserved))
      |          ~~~~~~~~~~~~~ ^~ ~~~~~~~~
      |          uint2_t {aka boost::multiprecision::number<boost::multiprecision::backends::cpp_int_backend<2, 2, boost::multiprecision::unsigned_magnitude, boost::multiprecision::unchecked, void>, boost::multiprecision::et_off>}
../../../../../../workspace/sources/phosphor-ipmi-host/user_channel/channelcommands.cpp:180:24: note: candidate: 'operator==(int, int)' (built-in)

+++ user_channel/channelcommands.cpp
@@ -177,7 +177,7 @@ ipmi ::RspType<uint3_t, // access mode,
         return responseInvalidFieldRequest();
     }
 
-    if ((accessSetMode == doNotSet) || (accessSetMode == reserved))
+    if ((accessSetMode == (int)doNotSet) || (accessSetMode == (int)reserved))
     {
         log<level::DEBUG>("Get channel access - Invalid Access mode");
         return responseInvalidFieldRequest();
@@ -197,11 +197,11 @@ ipmi ::RspType<uint3_t, // access mode,
 
     Cc compCode;
 
-    if (accessSetMode == nvData)
+    if (accessSetMode == (int)nvData)
     {
         compCode = getChannelAccessPersistData(chNum, chAccess);
     }
-    else if (accessSetMode == activeData)
+    else if (accessSetMode == (int)activeData)
     {
         compCode = getChannelAccessData(chNum, chAccess);
     }
